Moved CSV reader code to common code
